Correctly specify prereqs on MSWin32
On Windows, the Win32::TieRegistry module is used. This module is
not in core.

It's not specified in dist.ini and thus it's not in the Makefile.PL
generated by dzil.
The dist.ini does not have a 'native' way of determining prerequisites
that differ per OS platform, so we need to use the OSPrereqs plugin.