Fix POD type mismatch in item list
`podchecker` noticed a type mismatch in the item list altered in this
change.  The reason being that the first item, "-1", as an argument to
`=item` is interpreted by POD formatters as a definition section, whereas
the second item, "0", is interpreted as a number.  The solution presented
here removes the item list, reduces the amount of text and reflects a
similar text segment in the POD of the following section (the docs for
`openpage()`).  This change thus tries to remain consistent with the current
documentation style.