Use strict pragma wherever possible
... where every module in the distribution now uses the `strict` pragma.
Only in `PDF::API2` was it necessary to turn off strict for `'vars'`.  The
reason being the options passed around are sometimes handles as arrays and
other times as hashes and it would require lots of digging to sort that
particular issue out.  Nevertheless, it was possible to add `strict` to all
modules where this was missing and to make any variables lexical which
weren't already so.  In many cases the `new_api` method needed to have the
`$self` variable made lexical and this happened so often, and with the same
structure that it looked like this was a copy-paste issue.  Nevertheless,
this variable is now lexical inside `new_api` throughout the code.  The
tests all pass and the `Test::Kwalitee::Extra` tests no longer complain that
`use strict` isn't used throughout the distribution.  This change should
thus fix the `use_strict`
[CPANTS](http://cpants.cpanauthors.org/release/SSIMMS/PDF-API2-2.028)
violation and allows the distribution to use current Perl best practice.