Enable reading of numbers, booleans, and nulls in an object stream
Previously, there had to be an irregular character (e.g. space) after
each of these, in order for the parser to know that it was finished.
However, in an object stream, the value being read will be exactly a
number, true, false, or null, with no subsequent characters.  To
handle that case, allow them to match when at the end of the string.